+++ frontend/src/component/company/Company.js
@@ -6,13 +6,14 @@ import "./Company.css";
 class Company extends Component {
 
     render() {
+        let visibleCompany = this.props.companies.find(company => company.isVisible);
         let companyClassName = this.getCompanyClassName();
         let companyTextClassName = this.getCompanyTextClassName();
         return (
             <div className="company-container" onClick={this.props.onClick}>
                 <div className={companyClassName}>
                     <div className={companyTextClassName}>
-                        {this.props.visibleCompany}
+                        {visibleCompany.name}
                     </div>
                 </div>
             </div>
@@ -20,6 +21,10 @@ class Company extends Component {
     }
 
     getCompanyTextClassName() {
+        if (this.props.companies.length === 1) {
+            return "company__text";
+        }
+
         return this.props.chevronUp ?
             "company__text chevron chevron-up" :
             "company__text chevron chevron-down";
