+++ frontend/src/component/company/__list/CompanyList.js
@@ -1,4 +1,5 @@
 import {Component} from "react";
+import CompanyListItem from "../__list-item/CompanyListItem";
 
 import "./CompanyList.css";
 
@@ -6,11 +7,23 @@ class CompanyList extends Component {
 
     render() {
         let companyListClassName = this.getCompaniesClassName();
+        let companies = this.getCompaniesList();
         return (
             <ul className={companyListClassName}>
-                {this.props.content}
+                {companies}
             </ul>
         )
+    }
+
+    getCompaniesList() {
+        return this.props.companies
+            .filter(company => !company.isVisible)
+            .map((company) =>
+                <CompanyListItem key={company.id}
+                                 onClick={(e) => this.props.onClickItem(company.id, this.props.companies, e)}
+                                 content={company.name}
+                />
+            );
     }
 
     getCompaniesClassName() {
